requirements.txt
streamlit
pandas
numpy
pyarrow
openpyxl
pydeck
plotly
pyyaml
